Modify setup-keepassxc.sh

+++ setup-keepassxc.sh
@@ -3,15 +3,14 @@ set -eu
 
 repo_url=https://github.com/keepassxreboot/keepassxc
 
-install_fuse() {
+if [ "$(dpkg-query -f '${Status}' -W libfuse2 2>/dev/null)" != 'install ok installed' ]; then
   sudo apt install -y fuse libfuse2
   sudo modprobe fuse
   getent group fuse || sudo groupadd fuse
   sudo usermod -a -G fuse "$USER"
-}
+fi
 
-install_fuse
-sudo apt install -y curl
+[ "$(dpkg-query -f '${Status}' -W curl 2>/dev/null)" == 'install ok installed' ] || sudo apt install -y curl
 
 version=$(curl -sS -w '%{redirect_url}' -o /dev/null "$repo_url/releases/latest" | sed 's|.*/tag/||')
 
@@ -23,7 +22,7 @@ else
   app_filename=$(basename "$app_path")
   curl -Lo "$app_path" ${repo_url}/releases/download/${version}/${app_filename}
   chmod +x "$app_path"
-  echo Downloaded KeePassXC latest version ($version).
+  echo "Downloaded KeePassXC latest version ($version)".
 fi
 
 icon_path="$HOME/.icons/keepassxc-logo.svg"
